requirements.txt
requests
Django==1.9.6
django-jet==0.1.4
django-tastypie==0.13.3
django-wysiwyg-redactor==0.4.9
Pillow==3.0.0

